fix router structure
