+++ src/main.jsx
@@ -1,24 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.jsx";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import AppRouter from "./AppRouter.jsx";
 import "./index.css";
-
-const router = createBrowserRouter([
-  /*Para crear rutas una vez generado el .jsx dupliquen la estructura indicando el path y realicen la importación correspondiente
-  {
-    path: "/ejemplo",
-    element: <Ejemplo />,
-  },
-  */
-  {
-    path: "/",
-    element: <App />,
-  },
-]);
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <RouterProvider router={router} />
+    <AppRouter />
   </React.StrictMode>,
 );
