+++ src/pages/Login.jsx
@@ -1,6 +1,4 @@
-import "./App.css";
-
-function App() {
+function Login() {
   return (
     <>
       <div>Hola Alkemy!</div>
@@ -8,4 +6,4 @@ function App() {
   );
 }
 
-export default App;
+export default Login;
